requirements.txt
git-bob
openai
pygithub
